# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"Achievement": {
				"properties": {
					"title": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"title",
					"picture",
					"description",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AchievementsResponse": {
				"properties": {
					"achievements": {
						"items": {
							"$ref": "#/components/schemas/Achievement"
						},
						"type": "array"
					}
				},
				"required": [
					"achievements"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AchievementCreation": {
				"properties": {
					"title": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"persistent": {
						"type": "boolean"
					}
				},
				"required": [
					"title",
					"picture",
					"description",
					"persistent"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AchievementCreationRequest": {
				"properties": {
					"achievement": {
						"$ref": "#/components/schemas/AchievementCreation"
					}
				},
				"required": [
					"achievement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AchievementResponse": {
				"properties": {
					"achievement": {
						"$ref": "#/components/schemas/Achievement"
					}
				},
				"required": [
					"achievement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmptyDict": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Article": {
				"properties": {
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"published": {
						"type": "boolean"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"time_published": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "number",
						"format": "double"
					},
					"resource": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"body",
					"published",
					"year",
					"time_published",
					"picture",
					"id",
					"author",
					"resource"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DirContent": {
				"properties": {
					"id": {
						"type": "string"
					},
					"files": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"dirs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"files",
					"dirs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticlesResponse": {
				"properties": {
					"articles": {
						"items": {
							"$ref": "#/components/schemas/Article"
						},
						"type": "array"
					},
					"contents": {
						"items": {
							"$ref": "#/components/schemas/DirContent"
						},
						"type": "array"
					},
					"meta": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total"
						],
						"type": "object"
					}
				},
				"required": [
					"articles",
					"contents",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticleCreation": {
				"properties": {
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"published": {
						"type": "boolean"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"time_published": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"body",
					"published",
					"year",
					"time_published",
					"picture"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticleCreationRequest": {
				"properties": {
					"article": {
						"$ref": "#/components/schemas/ArticleCreation"
					}
				},
				"required": [
					"article"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ArticleResponse": {
				"properties": {
					"article": {
						"$ref": "#/components/schemas/Article"
					},
					"contents": {
						"items": {
							"$ref": "#/components/schemas/DirContent"
						},
						"type": "array"
					}
				},
				"required": [
					"article",
					"contents"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Thread": {
				"properties": {
					"public": {
						"type": "boolean"
					},
					"title": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"unread": {
						"type": "number",
						"format": "double"
					},
					"posts_count": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"public",
					"title",
					"year",
					"id",
					"unread",
					"posts_count",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ThreadsResponse": {
				"properties": {
					"threads": {
						"items": {
							"$ref": "#/components/schemas/Thread"
						},
						"type": "array"
					}
				},
				"required": [
					"threads"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreadResponse": {
				"properties": {
					"threads": {
						"$ref": "#/components/schemas/Thread"
					}
				},
				"required": [
					"threads"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreadCreation": {
				"properties": {
					"public": {
						"type": "boolean"
					},
					"title": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"public",
					"title",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreadsCreationRequest": {
				"properties": {
					"thread": {
						"$ref": "#/components/schemas/ThreadCreation"
					}
				},
				"required": [
					"thread"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"profile_picture": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"task_num": {
						"type": "number",
						"format": "double"
					},
					"achievements": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"enabled": {
						"type": "boolean"
					},
					"nick_name": {
						"type": "string"
					},
					"addr_country": {
						"type": "string"
					},
					"school_name": {
						"type": "string"
					},
					"seasons": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"successful": {
						"type": "boolean"
					},
					"cheat": {
						"type": "boolean"
					},
					"tasks": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"co_tasks": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"short_info": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"first_name",
					"last_name",
					"profile_picture",
					"gender",
					"role",
					"score",
					"task_num",
					"achievements",
					"enabled",
					"nick_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsersResponse": {
				"properties": {
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Wave": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"garant": {
						"type": "number",
						"format": "double"
					},
					"time_published": {
						"type": "string",
						"nullable": true
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"public": {
						"type": "boolean"
					},
					"sum_points": {
						"type": "number",
						"format": "double"
					},
					"task_cnt": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"index",
					"caption",
					"garant",
					"year",
					"id",
					"public",
					"sum_points",
					"task_cnt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Waves": {
				"properties": {
					"waves": {
						"items": {
							"$ref": "#/components/schemas/Wave"
						},
						"type": "array"
					}
				},
				"required": [
					"waves"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WaveResponse": {
				"properties": {
					"wave": {
						"$ref": "#/components/schemas/Wave"
					}
				},
				"required": [
					"wave"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WaveCreate": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"garant": {
						"type": "number",
						"format": "double"
					},
					"time_published": {
						"type": "string",
						"nullable": true
					},
					"year": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"index",
					"caption",
					"garant",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WaveCreationRequest": {
				"properties": {
					"wave": {
						"$ref": "#/components/schemas/WaveCreate"
					}
				},
				"required": [
					"wave"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WaveBase": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"garant": {
						"type": "number",
						"format": "double"
					},
					"time_published": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"index",
					"caption",
					"garant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WaveUpdateRequest": {
				"properties": {
					"wave": {
						"$ref": "#/components/schemas/WaveBase"
					}
				},
				"required": [
					"wave"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"ksi": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "https://ksi.fi.muni.cz/login/",
						"scopes": {}
					}
				}
			}
		}
	},
	"info": {
		"title": "web-backend-swagger",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "KSI"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/achievements": {
			"get": {
				"operationId": "AchievementsGetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AchievementsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "_limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "_start",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "achievement",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "AchievementsCreateNew",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AchievementCreationRequest"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AchievementCreationRequest"
							}
						}
					}
				}
			}
		},
		"/achievements/{achievementId}": {
			"get": {
				"operationId": "AchievementsGetSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AchievementResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "achievementId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AchievementsEditSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AchievementResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "achievementId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AchievementCreationRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "AchievementsDeleteSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmptyDict"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "achievementId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/articles": {
			"get": {
				"operationId": "ArticlesGetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticlesResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "_start",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "article",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "ArticlesCreateNew",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleCreationRequest"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ArticleCreationRequest"
							}
						}
					}
				}
			}
		},
		"/articles/{articleId}": {
			"get": {
				"operationId": "ArticlesGetSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "ArticlesEditSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ArticleCreationRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "ArticlesDeleteSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmptyDict"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/threads": {
			"get": {
				"operationId": "ThreadsGetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadsResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_wave",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "ThreadsCreateNew",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ThreadsCreationRequest"
							}
						}
					}
				}
			}
		},
		"/threads/{threadsId}": {
			"get": {
				"operationId": "ThreadsGetSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "threadsId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "ThreadsMarkVisited",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmptyDict"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "threadsId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "UsersGetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{usersId}": {
			"get": {
				"operationId": "UsersGetSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "usersId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "UsersDeleteSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmptyDict"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "usersId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/waves": {
			"get": {
				"operationId": "WavesGetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Waves"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "WavesCreateNew",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WaveResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WaveCreationRequest"
							}
						}
					}
				}
			}
		},
		"/waves/{waveId}": {
			"get": {
				"operationId": "WavesGetSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WaveResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "waveId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "WavesUpdateSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WaveResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "waveId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WaveUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "WavesDeleteSingle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmptyDict"
								}
							}
						}
					}
				},
				"security": [
					{
						"ksi": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "waveId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}